+++ app/src/main/java/com/adit/bangkit/plagroid/ui/activities/AddProductActivity.kt
@@ -238,7 +238,7 @@ class AddProductActivity : BaseActivity(), View.OnClickListener {
             et_product_title.text.toString().trim { it <= ' ' },
             et_product_price.text.toString().trim { it <= ' ' }.toInt(),
             et_product_description.text.toString().trim { it <= ' ' },
-            et_product_quantity.text.toString().trim { it <= ' ' }.toInt(),
+            et_product_quantity.text.toString().trim { it <= ' ' },
             mProductImageURL
         )
